chore: workflow failure fix
