+++ scripts/soak-test-kubectl.test.ts
@@ -18,8 +18,11 @@ import {
   checkPodStability,
   collectMetrics,
   failWithReason,
+  assertMetrics,
+  assertCacheMissGrowth,
   SoakTestFailure,
 } from "./soak-test.js";
+import { STABILIZATION_ITERATIONS } from "./soak-constants.js";
 
 const mockExecSync = vi.mocked(execSync);
 
@@ -194,5 +197,119 @@ describe("checkPodStability", () => {
 
     const initialPods = new Set(["pod-a"]);
     expect(() => checkPodStability(initialPods, 5, tmpDir)).not.toThrow();
+  });
+});
+
+describe("assertMetrics", () => {
+  let tmpDir: string;
+
+  beforeEach(() => {
+    mockExecSync.mockReset();
+    tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), "soak-assert-"));
+  });
+
+  afterEach(() => {
+    fs.rmSync(tmpDir, { recursive: true, force: true });
+  });
+
+  function writeCsv(rows: string[]): void {
+    const header =
+      "iteration,timestamp,watch_controller_failures_delta,pepr_cache_miss_delta,pepr_resync_failure_count";
+    fs.writeFileSync(path.join(tmpDir, "metrics.csv"), [header, ...rows].join("\n") + "\n");
+  }
+
+  it("passes when all metrics are zero", () => {
+    writeCsv(["1,2024-01-01T00:00:00Z,0,0,0"]);
+    expect(() => assertMetrics(1, tmpDir)).not.toThrow();
+  });
+
+  it("fails when watch controller failures are non-zero", () => {
+    writeCsv(["1,2024-01-01T00:00:00Z,3,0,0"]);
+    expect(() => assertMetrics(1, tmpDir)).toThrow("Watch controller failures detected: 3");
+  });
+
+  it("fails when resync failures exceed threshold", () => {
+    writeCsv(["1,2024-01-01T00:00:00Z,0,0,10"]);
+    expect(() => assertMetrics(1, tmpDir, 5)).toThrow("Resync failures exceeded threshold: 10 > 5");
+  });
+
+  it("does not check cache miss growth before stabilization", () => {
+    writeCsv(["1,2024-01-01T00:00:00Z,0,100,0"]);
+    // iteration <= STABILIZATION_ITERATIONS, so cache miss growth is not checked
+    expect(() => assertMetrics(1, tmpDir)).not.toThrow();
+  });
+
+  it("checks cache miss growth after stabilization", () => {
+    const rows: string[] = [];
+    for (let i = 1; i <= STABILIZATION_ITERATIONS + 1; i++) {
+      const cacheDelta = i === STABILIZATION_ITERATIONS ? 2 : 0;
+      rows.push(`${i},2024-01-01T00:00:00Z,0,${cacheDelta},0`);
+    }
+    rows.push(`${STABILIZATION_ITERATIONS + 2},2024-01-01T00:00:00Z,0,50,0`);
+    writeCsv(rows);
+
+    expect(() => assertMetrics(STABILIZATION_ITERATIONS + 2, tmpDir, 5, 10)).toThrow(
+      "Cache misses grew",
+    );
+  });
+
+  it("passes cache miss growth check when within threshold", () => {
+    const rows: string[] = [];
+    for (let i = 1; i <= STABILIZATION_ITERATIONS + 1; i++) {
+      const cacheDelta = i === STABILIZATION_ITERATIONS ? 5 : 0;
+      rows.push(`${i},2024-01-01T00:00:00Z,0,${cacheDelta},0`);
+    }
+    rows.push(`${STABILIZATION_ITERATIONS + 2},2024-01-01T00:00:00Z,0,8,0`);
+    writeCsv(rows);
+
+    expect(() => assertMetrics(STABILIZATION_ITERATIONS + 2, tmpDir, 5, 10)).not.toThrow();
+  });
+});
+
+describe("assertCacheMissGrowth", () => {
+  let tmpDir: string;
+
+  beforeEach(() => {
+    mockExecSync.mockReset();
+    tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), "soak-growth-"));
+  });
+
+  afterEach(() => {
+    fs.rmSync(tmpDir, { recursive: true, force: true });
+  });
+
+  function writeCsv(rows: string[]): void {
+    const header =
+      "iteration,timestamp,watch_controller_failures_delta,pepr_cache_miss_delta,pepr_resync_failure_count";
+    fs.writeFileSync(path.join(tmpDir, "metrics.csv"), [header, ...rows].join("\n") + "\n");
+  }
+
+  it("skips check when not enough CSV rows", () => {
+    writeCsv(["1,2024-01-01T00:00:00Z,0,5,0"]);
+    expect(() => assertCacheMissGrowth(100, tmpDir, 10)).not.toThrow();
+  });
+
+  it("throws when growth exceeds threshold", () => {
+    const rows: string[] = [];
+    for (let i = 1; i <= STABILIZATION_ITERATIONS; i++) {
+      const cacheDelta = i === STABILIZATION_ITERATIONS ? 3 : 0;
+      rows.push(`${i},2024-01-01T00:00:00Z,0,${cacheDelta},0`);
+    }
+    writeCsv(rows);
+
+    expect(() => assertCacheMissGrowth(20, tmpDir, 10)).toThrow(
+      "Cache misses grew from 3 to 20 (growth: 17 > threshold: 10)",
+    );
+  });
+
+  it("passes when growth is within threshold", () => {
+    const rows: string[] = [];
+    for (let i = 1; i <= STABILIZATION_ITERATIONS; i++) {
+      const cacheDelta = i === STABILIZATION_ITERATIONS ? 3 : 0;
+      rows.push(`${i},2024-01-01T00:00:00Z,0,${cacheDelta},0`);
+    }
+    writeCsv(rows);
+
+    expect(() => assertCacheMissGrowth(8, tmpDir, 10)).not.toThrow();
   });
 });
